plumbing: rename pack session
